+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 # 📋 Changelog for v1.0.4
 
+- Pressing "clear text" icon in Search box on search page, brings back keyboard focus on Search box
+-